Fixed autoStack interrupt
fixed interrupt logic error but also removed it from the thread itself in favor of join scope termination

(if not automatic, two options for thread interrupts possible)

in Blue4 and Red4 I quick fixed some sleep commands that were task specific, not thread specific.  Still not sure if it matters, but best be safe.